Changed Recordings.dart to SearchPage.dart
